+++ LockSecurity/src/me/choco/locks/commands/MainCommand.java
@@ -4,6 +4,7 @@ import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
+import org.bukkit.configuration.file.YamlConfiguration;
 
 import me.choco.locks.LockSecurity;
 
@@ -15,6 +16,8 @@ public class MainCommand implements CommandExecutor{
 	
 	@Override
 	public boolean onCommand(CommandSender sender, Command cmd, String commandLabel, String[] args){
+		@SuppressWarnings("deprecation")
+		YamlConfiguration pluginFile = YamlConfiguration.loadConfiguration(getClass().getResourceAsStream("/plugin.yml"));
 		if (args.length == 1){
 			if (args[0].equalsIgnoreCase("reload")){
 				if (sender.hasPermission("locks.reload")){
@@ -30,7 +33,7 @@ public class MainCommand implements CommandExecutor{
 			if (args[0].equalsIgnoreCase("version")){
 				sender.sendMessage(ChatColor.GOLD + "--------------------------------------------");
 				sender.sendMessage(ChatColor.DARK_AQUA + "" + ChatColor.BOLD + "Version: " + ChatColor.RESET + ChatColor.GRAY  + plugin.getDescription().getVersion());
-				sender.sendMessage(ChatColor.DARK_AQUA + "" + ChatColor.BOLD + "API Version: " + ChatColor.RESET + ChatColor.GRAY + "LS-1.0A_v1.8.8");
+				sender.sendMessage(ChatColor.DARK_AQUA + "" + ChatColor.BOLD + "API Version: " + ChatColor.RESET + ChatColor.GRAY + pluginFile.getString("apiversion"));
 				sender.sendMessage(ChatColor.DARK_AQUA + "" + ChatColor.BOLD + "Developer / Maintainer: " + ChatColor.RESET + ChatColor.GRAY + "2008Choco");
 				sender.sendMessage(ChatColor.DARK_AQUA + "" + ChatColor.BOLD + "Development Page: " + ChatColor.RESET + ChatColor.GRAY + "http://dev.bukkit.org/bukkit-plugins/lock-security");
 				sender.sendMessage(ChatColor.DARK_AQUA + "" + ChatColor.BOLD + "Report Bugs To: " + ChatColor.RESET + ChatColor.GRAY + "http://dev.bukkit.org/bukkit-plugins/lock-security/tickets");